project-bolt-secret-python-ybpuunrz3.0/project/requirements.txt
okx-python>=0.4.0
pandas>=1.3.0
python-dotenv>=0.19.0
numpy>=1.21.0
python-dateutil>=2.8.2